+++ gulpfile.js
@@ -37,7 +37,13 @@ build.rig.getTasks = function () {
   return result;
 };
 
-build.initialize(gulp);
+/* fast-serve */
+const { addFastServe } = require("spfx-fast-serve-helpers");
+addFastServe(build);
+/* end of fast-serve */
+
+build.initialize(require('gulp'));
+
 
 const karmaTask = build.karma;
 if (karmaTask) {