+++ src/controls/dynamicForm/IDynamicFormState.ts
@@ -1,0 +1,9 @@
+
+import { IDynamicFieldProps } from './dynamicField/IDynamicFieldProps';
+export interface IDynamicFormState {
+  fieldCollection: IDynamicFieldProps[];
+  isSaving?: boolean;
+}
+
+
+