+++ src/controls/dynamicForm/dynamicField/IDynamicFieldState.ts
@@ -1,0 +1,6 @@
+export interface IDynamicFieldState {
+  /**
+  * The options available to the listPicker
+  */
+  changedValue: any;
+}