+++ src/controls/filePicker/FilePicker.tsx
@@ -212,6 +212,7 @@ export class FilePicker extends React.Component<
               <SiteFilePickerTab
                 fileBrowserService={this.fileBrowserService}
                 includePageLibraries={this.props.includePageLibraries}
+                defaultFolderAbsolutePath={this.props.defaultFolderAbsolutePath}
                 {...linkTabProps}
               />
             )}
@@ -359,7 +360,7 @@ export class FilePicker extends React.Component<
       this.state.organisationAssetsEnabled
     ) {
       links.push({
-        name: "Your organisation",
+        name: strings.OrgAssetsLinkLabel,
         url: addUrl ? "#orgAssets" : undefined,
         icon: "FabricFolderConfirm",
         key: "keyOrgAssets",