+++ src/loc/en-us.ts
@@ -2,6 +2,9 @@ declare var define: any;
 
 define([], () => {
   return {
+    "Save": "Save",
+    "Cancel": "Cancel",
+
     'SiteBreadcrumbLabel': 'Website breadcrumb',
 
     'ListViewGroupEmptyLabel': 'Empty',
@@ -310,15 +313,15 @@ define([], () => {
     SizeOptionMedium: "Medium",
     SizeOptionSmall: "Small",
     SizeUnit: [
-     "bytes",
-     "KB",
-     "MB",
-     "GB",
-     "TB",
-     "PB",
-     "EB",
-     "ZB",
-     "YB"
+      "bytes",
+      "KB",
+      "MB",
+      "GB",
+      "TB",
+      "PB",
+      "EB",
+      "ZB",
+      "YB"
     ],
     SorryWebSearch: "Sorry, this function isn't implemented in this sample, because it would require a Bing API key.",
     SortedAscending: "Sorted A to Z",
@@ -354,21 +357,31 @@ define([], () => {
     CollectionDataItemMissingFields: "Fields collection is empty!",
     InvalidUrlError: "The provided URL is not valid",
     CollectionDataSearch: "Search",
-MyTeamsLoadingMessage: "loading your teams",
-MyTeamsMessageError: "Something went wrong while loading your teams, please try later or refresh browser",
+    MyTeamsLoadingMessage: "loading your teams",
+    MyTeamsMessageError: "Something went wrong while loading your teams, please try later or refresh browser",
 
-MyTeamsNoTeamsMessage: "You don't have any teams",
-MyTeamsTeamChannelPublicMessage: "Public Channels",
-MyTeamsTeamChannelTypeMessage: "Private Channels",
-TeamChannelPickerFontIconFavoriteText: "Favorite",
-TeamChannelPickerFontIconPrivateChannelTitle: "Private Channel",
-TeamChannelPickerSugestionHeaderText: "Suggested Team Channels",
-TeamPickerButtonRemoveTitle: "remove",
-TeamPickernoResultsFoundText: "No Teams found",
-TeamPickerSugestionsHeaderText: "Suggested Teams",
-TeamsChannelPickerButtonRemoveTitle: "remove",
-TeamsChannelPickerNoresultsFoundText: "No channels found",
+    MyTeamsNoTeamsMessage: "You don't have any teams",
+    MyTeamsTeamChannelPublicMessage: "Public Channels",
+    MyTeamsTeamChannelTypeMessage: "Private Channels",
+    TeamChannelPickerFontIconFavoriteText: "Favorite",
+    TeamChannelPickerFontIconPrivateChannelTitle: "Private Channel",
+    TeamChannelPickerSugestionHeaderText: "Suggested Team Channels",
+    TeamPickerButtonRemoveTitle: "remove",
+    TeamPickernoResultsFoundText: "No Teams found",
+    TeamPickerSugestionsHeaderText: "Suggested Teams",
+    TeamsChannelPickerButtonRemoveTitle: "remove",
+    TeamsChannelPickerNoresultsFoundText: "No channels found",
 
-    ViewMore: "View more"
+    ViewMore: "View more",
+
+    DynamicFormLoading: "Loading...",
+    DynamicFormPleaseWait: "Please wait...",
+    DynamicFormRequiredErrorMessage: "You can't leave this blank.",
+    DynamicFormTermPanelTitle: "Select Term",
+    DynamicFormEnterURLPlaceholder: "Enter a URL",
+    DynamicFormEnterDescriptionPlaceholder: "Alternative text",
+
+    customDisplayName: "Use this location:",
+    OrgAssetsLinkLabel: "Your organisation"
   };
 });