+++ src/loc/mystrings.d.ts
@@ -15,6 +15,9 @@ declare interface IControlStrings {
   TeamChannelPickerFontIconPrivateChannelTitle: string;
   TeamChannelPickerFontIconFavoriteText: string;
   TeamsChannelPickerButtonRemoveTitle: string;
+  // common
+  Save: string;
+  Cancel: string;
 
   PeoplePickerGroupNotFound: string;
   ListViewFilterLabel: string;
@@ -312,6 +315,7 @@ declare interface IControlStrings {
   StockImagesLinkLabel: string;
   StockImagesHeader: string;
   Yes: string;
+  OrgAssetsLinkLabel: string;
 
   // folder explorer
   NewFolderNamePlaceholder: string;
@@ -342,6 +346,17 @@ declare interface IControlStrings {
   CollectionDataSearch: string;
 
   ViewMore: string;
+
+  // Dynamic Form
+  DynamicFormLoading: string;
+  DynamicFormPleaseWait: string;
+  DynamicFormRequiredErrorMessage: string;
+  DynamicFormTermPanelTitle: string;
+  DynamicFormEnterURLPlaceholder: string;
+  DynamicFormEnterDescriptionPlaceholder: string;
+
+  // Location picker
+  customDisplayName:string;
 }
 
 declare interface IDateTimeStrings {